pyproject.toml
[project]
name = "personal-website"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2",
    "django-ckeditor>=6.7.2",
    "django-ckeditor-5>=0.2.17",
    "django-tinymce>=4.1.0",
    "dotenv>=0.9.9",
]
